+++ Sources/LiveKit/Extensions/Primitives.swift
@@ -41,3 +41,11 @@ extension URL {
         scheme == "https" || scheme == "wss"
     }
 }
+
+extension Double {
+
+    internal func rounded(to places: Int) -> Double {
+        let divisor = pow(10.0, Double(places))
+        return (self * divisor).rounded() / divisor
+    }
+}
